Get recipe images working (finally!!)
