+++ src/App.tsx
@@ -3,23 +3,8 @@ import { Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navbar";
 import RecipeGrid from "./components/RecipeGrid";
 import RecipeDetails from "./components/RecipeDetails";
-import { collection, getDocs } from "firebase/firestore";
-import { db } from "./constants/firebase";
-import { getIngredientDetails } from "./constants/utils";
-import axios, {AxiosResponse} from "axios";
+import { getIngredients, search } from "./constants/api";
 import "./stylesheets/App.scss";
-
-const SEARCH_ENDPOINT = "https://us-central1-recipe--finder.cloudfunctions.net/searchRecipes";
-
-type SearchResponse = {
-  page: {
-    current: number,
-    total_pages: number,
-    total_results: number,
-    size: number
-  },
-  results: SearchedRecipe[]
-};
 
 const App = () => {
   const [searchValue, setSearchValue] = React.useState("");
@@ -30,24 +15,13 @@ const App = () => {
 
   // fetch all ingredients from Firebase
   React.useEffect(() => {
-    const options: IngredientOption[] = [];
-    getDocs(collection(db, "ingredients")).then((querySnapshot) => {
-      querySnapshot.forEach((doc) => {
-        options.push({ id: doc.id, name: doc.data().name });
-      });
-      setIngredientOptions(options);
-    });
+    getIngredients().then(setIngredientOptions);
   }, []);
 
+  // execute a search query
   React.useEffect(() => {
-    axios.post(SEARCH_ENDPOINT, {
-      text: "",
-      ingredients: queriedIngredients.map(({ id }) => getIngredientDetails(id, ingredientOptions)?.name)
-        .filter((ingredient) => !!ingredient)
-    }).then(({ data }: AxiosResponse<SearchResponse>) => {
-      setRecipeMatches(data.results);
-      setNextPage(data.page.current === data.page.total_pages ? null : data.page.current + 1);
-    });
+    search(queriedIngredients, ingredientOptions, setNextPage)
+      .then((results) => setRecipeMatches(results));
   }, [ingredientOptions, queriedIngredients])
 
   return (
